add commit

+++ src/main/java/com/tactfactory/demo/dtos/BookDto.java
@@ -4,23 +4,27 @@ public class BookDto {
 	
 	private String name;
     private Integer nbPage;
-    private Integer price;
+    private Float price;
+    
+    
 	public String getName() {
 		return name;
 	}
 	public void setName(String name) {
 		this.name = name;
 	}
+	
 	public Integer getNbPage() {
 		return nbPage;
 	}
 	public void setNbPage(Integer nbPage) {
 		this.nbPage = nbPage;
 	}
-	public Integer getPrice() {
+	
+	public Float getPrice() {
 		return price;
 	}
-	public void setPrice(Integer price) {
+	public void setPrice(Float price) {
 		this.price = price;
 	}
